+++ apps/api/src/modules/sales/sales.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { AdminSalesController, PublicSalesController } from './sales.controller';
-import { SalesService } from '@sitekit/core';
+import { SalesService, SalesTemplateService } from '@sitekit/core';
 
-@Module({ controllers: [PublicSalesController, AdminSalesController], providers: [SalesService], exports: [SalesService] })
+@Module({ controllers: [PublicSalesController, AdminSalesController], providers: [SalesService, SalesTemplateService], exports: [SalesService, SalesTemplateService] })
 export class SalesModule {}